requirements.txt
Flask==2.0.2
Flask-Cors==3.0.10
python-docx==0.8.11
PyPDF2==1.26.0
gunicorn==20.1.0